+++ src/IdeatecAPI.Application/Features/Productos/DTO/EditarProductoDTO.cs
@@ -24,7 +24,7 @@ public class EditarProductoDTO
     // Datos de SucursalProducto
     public int SucursalProductoId { get; set; }
     public decimal? PrecioUnitario { get; set; }
-    public int? Stock { get; set; }
+    public decimal? Stock { get; set; }
     public decimal? PrecioMayorista { get; set; }
     public int? CantidadMinimaMayorista { get; set; }
     public bool? EnPromocion { get; set; }
