+++ src/IdeatecAPI.Application/Features/Proveedor/DTOs/ObtenerCompraProveedorDTO.cs
@@ -10,7 +10,7 @@ public class ObtenerCompraProveedorDTO
     public int ProductoId { get; set; }
     public string? NomProducto { get; set; }
     public decimal? PrecioCompra { get; set; }
-    public int? Cantidad { get; set; }
+    public decimal? Cantidad { get; set; }
     public string? UnidadMedida { get; set; }
     public string? DocReferencia { get; set; }
     public DateTime? FechaCreacion { get; set; }
